@@ -1,11 +1,11 @@
 name: ios_device_check
 description: Flutter plugin for accessing the iOS DeviceCheck. Only supported on iOS 11.0 or newer.
-version: 1.3.0+1
+version: 2.0.0
 homepage: https://github.com/okadan/flutter-ios-device-check
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0 <2.0.0"
+  sdk: ">=2.12.0 <3.0.0"
+  flutter: ">=1.20.0"
 
 dependencies:
   flutter:
